Added bin folder and moved exercise_bot_ruby.rb file to that folder
