+++ bin/exercise_bot_ruby.rb
