+++ spec/slack_exercisebot/commands/exercise_spec.rb
@@ -8,7 +8,6 @@ describe SlackExercisebot::Commands::GetExercise do
   subject { app }
 
   it 'returns exercise #1' do
-    expect(message: "#{SlackRubyBot.config.user} Exercise1", channel: 'channel').to
-    respond_with_slack_message('Lets start with High Knees. Do it for 20 seconds!!')
+    expect(message: "#{SlackRubyBot.config.user} Exercise1", channel: 'channel').to respond_with_slack_message('Lets start with High Knees. Do it for 20 seconds!!')
   end
 end
