+++ spec/spec_helper.rb
@@ -1,4 +1,4 @@
 $LOAD_PATH.unshift(File.join(File.dirname(__FILE__), '..'))
 
 require 'slack-ruby-bot/rspec'
-require 'exercise_bot_ruby'
+require 'bin/exercise_bot_ruby'
